Book : Repository, Controller, Service, Model OK

+++ src/main/java/com/bibliomanager/library/repository/EditorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
  @Repository: repository in the persistence layer and makes it eligible for Spring’s exception translation mechanism.
 
 */
-public interface EditorRepository extends CrudRepository<Editor, Integer> {
+public interface EditorRepository extends CrudRepository<Editor, Long> {
 
 }
