+++ src/main/java/com/bibliomanager/library/repository/TypeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
  @Repository: repository in the persistence layer and makes it eligible for Spring’s exception translation mechanism.
 
 */
-public interface TypeRepository extends CrudRepository<Type, Integer> {
+public interface TypeRepository extends CrudRepository<Type, Long> {
 
 }
